// src/web/models/DeviceCategories.ts
export const enum DEVICE_CATEGORIES {
    ROUTER = "router",
    ROUTER_WIFI = "router_wifi",
    PHONE = "phone",
    LAPTOP = "laptop",
    WORKSTATION = "workstation",
    TABLET = "tablet",
    TV = "tv",
    HUB = "hub",
    CAMERA = "camera",
    SMART_SPEAKER = "smart_speaker",
    UNKNOWN = "unknown",
}